Extract room tunneler
