+++ src/ship-gen/shipGenerator.js
@@ -119,7 +119,6 @@ export class Ship {
             this.rooms.push(bridge);
         }
 
-
         for (var i = 1; i < this.rooms.length; i++) {
             this.placeEntitiesInRoom(this.rooms[i]);
         }
@@ -155,33 +154,12 @@ export class Ship {
                     this.gameMap.floorTiles[x][y] = Tiles.lightFloor(x, y);
                 }
             }
-
-            if (i == 0) {
-                this.gameMap.wallTiles[newRoom.x1 + 1][newRoom.y1 + 1] = Tiles.redTorch(newRoom.x1 + 1, newRoom.y1 + 1);
-                this.gameMap.wallTiles[newRoom.x2 - 2][newRoom.y1 + 1] = Tiles.yellowTorch(newRoom.x2 - 2, newRoom.y1 + 1);
-            }
-
-            if (i !== 0) {
-                var lastRoom = this.rooms[this.rooms.length - 1];
-                var lastRoomCenter = lastRoom.center();
-                var newRoomCenter = newRoom.center();
-    
-                this._tunnelBetween(lastRoomCenter.x, lastRoomCenter.y, newRoomCenter.x, newRoomCenter.y);
-    
-                this.placeEntitiesInRoom(newRoom);
-            }
     
             this.rooms.push(newRoom);
         }
 
         console.log('Created room: ' + newRoom);
         return newRoom;
-    }
-
-    _tunnelBetweenRooms(room1, room2) {
-        var lastRoomCenter = room1.center();
-        var newRoomCenter = room2.center();
-        this._tunnelBetween(lastRoomCenter.x, lastRoomCenter.y, newRoomCenter.x, newRoomCenter.y);
     }
 
     // Sets the player coordinates based on the first room.
@@ -214,48 +192,6 @@ export class Ship {
                     EntityFactories.automatedTurret.spawn(this.gameMap, x, y);
                 }
             }
-        }
-    }
-
-    _tunnelBetween(x1, y1, x2, y2) {
-        if (Srand.intInRange(0, 1) == 1) {
-            // horizontal first, then vertical
-            this._createTunnel(x1, x2, y1, true);
-            this._createTunnel(y1, y2, x2, false);
-        } else {
-            // vertical first, then horizontal
-            this._createTunnel(y1, y2, x1, false);
-            this._createTunnel(x1, x2, y2, true);
-        }
-    }
-
-    _createTunnel(axisStart, axisEnd, otherAxis, isHorizontal) {
-        const start = Math.min(axisStart, axisEnd);
-        const end =  Math.max(axisStart, axisEnd);
-        for (var axisCoord = start; axisCoord <= end; axisCoord++) {
-            var x = isHorizontal ? axisCoord : otherAxis;
-            var y = isHorizontal ? otherAxis : axisCoord;
-
-            this.gameMap.floorTiles[x][y] = Tiles.lightFloor(x, y);
-            this.gameMap.wallTiles[x][y] = null; // remove any wall
-    
-            var xCheckTile1 = isHorizontal ? x : x - 1;
-            var yCheckTile1 = isHorizontal ? y - 1 : y;
-            this._tunnelAdjacent(xCheckTile1, yCheckTile1);
-    
-            var xCheckTile2 = isHorizontal ? x : x + 1;
-            var yCheckTile2 = isHorizontal ? y + 1 : y;
-            this._tunnelAdjacent(xCheckTile2, yCheckTile2);
-        }
-    }
-
-    _tunnelAdjacent(x, y) {
-        // do not place a wall if there is already an open floor here
-        var floorTileAdjacent = this.gameMap.floorTiles[x][y];
-        var wallTileAdjacent = this.gameMap.wallTiles[x][y];
-        if (!floorTileAdjacent && !wallTileAdjacent) {
-            this.gameMap.wallTiles[x][y] = Tiles.wall(x, y);
-            this.gameMap.floorTiles[x][y] = Tiles.darkFloor(x, y);
         }
     }
 }
